+++ src/JInterval.js
@@ -15,12 +15,12 @@ JInterval.getComma = require('./api/class/getComma')
 
 // General
 JInterval.prototype.copy = require('./api/general/copy')
-JInterval.prototype.getAlg = require('./api/general/getAlg')
+JInterval.prototype.getAlg = require('./api/notation/getAlg')
 JInterval.prototype.getFraction = require('./api/general/getFraction')
 JInterval.prototype.getPeo = require('./api/general/getPeo')
 JInterval.prototype.getPitch = require('./api/general/getPitch')
 JInterval.prototype.getPitchClass = require('./api/general/getPitchClass')
-JInterval.prototype.getVal = require('./api/general/getVal')
+JInterval.prototype.getAsDecimal = require('./api/general/getAsDecimal')
 JInterval.prototype.toString = require('./api/general/toString')
 
 // Notation
@@ -29,6 +29,7 @@ JInterval.prototype.getEndNotation = require('./api/notation/getEndNotation')
 // Frequency
 JInterval.prototype.getEndFreqHz = require('./api/freq/getEndFreqHz')
 JInterval.prototype.getEndFreqText = require('./api/freq/getEndFreqText')
+JInterval.prototype.getStartFreqHz = require('./api/freq/getStartFreqHz')
 
 // Maths
 JInterval.prototype.get1 = require('./api/maths/get1')
