+++ src/api/general/getAsDecimal.js
@@ -1,0 +1,5 @@
+var getAsDecimal = function() {
+  return this.peo.getVal()
+}
+
+module.exports = getAsDecimal
