--- src/api/general/getVal.js
@@ -1,5 +1,0 @@
-var getVal = function() {
-  return this.peo.getVal()
-}
-
-module.exports = getVal
