+++ src/api/maths/get1.js
@@ -1,4 +1,4 @@
-// Return an identity JInterval, representing a JInterval on 1/1
+// Return an identity JInterval, representing an interval of unison, 1/1
 var get1 = function() {
   var JIntervalConstructor = this.constructor
   return new JIntervalConstructor()
