+++ src/api/maths/mult.js
@@ -1,13 +1,13 @@
+// Multiply one JInterval by another, optionally to a power (repeated that many times)
 var mult = function(otherJInterval, power) {
-
   // Use Peo to do a multiplication
-  var peo1 = this.peo
-  var peo2 = otherJInterval.peo
-  var peoMult = peo1.mult(peo2, power)
+  var thisPeo = this.peo
+  var thatPeo = otherJInterval.peo
+  var peoMult = thisPeo.mult(thatPeo, power)
+  // Use the constructor to make a new JInterval based on peoMult
   var JIntervalConstructor = this.constructor
   var newJInterval = new JIntervalConstructor(peoMult)
   return newJInterval
-
 }
 
 module.exports = mult
