+++ src/api/maths/pow.js
@@ -1,12 +1,12 @@
+// Put a JInterval to an integer power, which means repeat it that many times.
 var pow = function(power) {
-
   // Use Peo to provide a power
-  var peo = this.peo
-  var peoPower = peo.pow(power)
+  var thisPeo = this.peo
+  var newPeo = thisPeo.pow(power)
+  // Use the constructor to make a new JInterval based on newPeo
   var JIntervalConstructor = this.constructor
-  var newJInterval = new JIntervalConstructor(peoPower)
+  var newJInterval = new JIntervalConstructor(newPeo)
   return newJInterval
-
 }
 
 module.exports = pow
